"check_style_score -> 4 of 5 passed. style score: 0/1 "

+++ Mod_2/Assignment-3/Solution.java
@@ -8,13 +8,13 @@ public class Solution {
         Scanner s=new Scanner(System.in);      
         int base = s.nextInt();
         int exponent = s.nextInt();
-        float result=power(base,exponent);
+        int result=power(base,exponent);
         System.out.println(result);
     }
     /*
     Need to write the power function and print the output.
     */
-    public static float power(final int b, final int e){
+    public static int power(final int b, final int e){
         if (e > 0) {
             return b * (power(b, e-1));
         } else {
